backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
PyMySQL==1.1.0
Pillow>=10.2.0
django-filter==23.5
python-decouple==3.8
djangorestframework-simplejwt==5.3.1
pandas>=2.2.0
python-dateutil==2.8.2
